Move MessageQueueState class to its own file
... so that I can use it from two places.
