+++ main/lsp/LSPPreprocessor.h
@@ -4,6 +4,7 @@
 #include "absl/synchronization/mutex.h"
 #include "main/lsp/LSPConfiguration.h"
 #include "main/lsp/LSPMessage.h"
+#include "main/lsp/MessageQueueState.h"
 #include <deque>
 
 namespace sorbet::realmain::lsp {
@@ -15,26 +16,6 @@ class DidCloseTextDocumentParams;
 class DidOpenTextDocumentParams;
 class WatchmanQueryResponse;
 class CancelParams;
-
-struct MessageQueueState {
-    std::deque<std::unique_ptr<LSPMessage>> pendingRequests;
-    bool terminate = false;
-    int errorCode = 0;
-    // Counters collected from other threads.
-    CounterState counters;
-
-    class NotifyOnDestruction {
-        absl::Mutex &mutex;
-        bool &flag;
-
-    public:
-        NotifyOnDestruction(MessageQueueState &state, absl::Mutex &mutex) : mutex(mutex), flag(state.terminate){};
-        ~NotifyOnDestruction() {
-            absl::MutexLock lck(&mutex);
-            flag = true;
-        }
-    };
-};
 
 class TaskQueue final {
     absl::Mutex stateMutex;
